refactor(opentitanlib): resolve internal circular dependencies
Resolve circular dependencies within `opentitanlib` to prepare for modularization (Phase 1).

Specifically:
- Move generic polling helper to `util::poll`.
- Move transport capabilities to `io::capabilities`.
- Move serializable error traits to `io::serializable_error` and transport errors to `io::errors`.
- Move bootstrap type parameters to `io::bootstrap_types`.
- Refactor parameter-based target creation from `Params` structs in `io` to `TransportWrapper` methods in `app`.
- Update all callers in `opentitanlib` and `opentitantool` to use the new APIs.

TAG=agy
CONV=b1350f87-77e1-4940-8d1e-55a68667743e
