+++ sw/host/opentitanlib/BUILD
@@ -109,15 +109,19 @@ rust_library(
         "src/image/manifest_def.rs",
         "src/image/manifest_ext.rs",
         "src/image/mod.rs",
+        "src/io/bootstrap_types.rs",
+        "src/io/capabilities.rs",
         "src/io/console.rs",
         "src/io/eeprom.rs",
         "src/io/emu.rs",
+        "src/io/errors.rs",
         "src/io/gpio.rs",
         "src/io/i2c.rs",
         "src/io/ioexpander.rs",
         "src/io/jtag.rs",
         "src/io/mod.rs",
         "src/io/nonblocking_help.rs",
+        "src/io/serializable_error.rs",
         "src/io/spi.rs",
         "src/io/uart.rs",
         "src/io/usb/desc.rs",
@@ -195,7 +199,6 @@ rust_library(
         "src/transport/dediprog/gpio.rs",
         "src/transport/dediprog/mod.rs",
         "src/transport/dediprog/spi.rs",
-        "src/transport/errors.rs",
         "src/transport/ftdi/chip.rs",
         "src/transport/ftdi/gpio.rs",
         "src/transport/ftdi/mod.rs",
@@ -253,6 +256,7 @@ rust_library(
         "src/util/mod.rs",
         "src/util/num_de.rs",
         "src/util/parse_int.rs",
+        "src/util/poll.rs",
         "src/util/present.rs",
         "src/util/printer.rs",
         "src/util/raw_tty.rs",
