+++ sw/host/opentitanlib/src/app/mod.rs
@@ -1037,6 +1037,57 @@ impl TransportWrapper {
 
         Ok(())
     }
+
+    pub fn create_uart(&self, params: &crate::io::uart::UartParams) -> Result<Rc<dyn Uart>> {
+        let uart = self.uart(&params.uart)?;
+        if let Some(baudrate) = params.baudrate {
+            uart.set_baudrate(baudrate)?;
+        }
+        log::info!("set_flow_control to {}", params.flow_control);
+        uart.set_flow_control(params.flow_control)?;
+        Ok(uart)
+    }
+
+    pub fn create_spi(
+        &self,
+        params: &crate::io::spi::SpiParams,
+        default_instance: &str,
+    ) -> Result<Rc<dyn Target>> {
+        let spi = self.spi(params.bus.as_deref().unwrap_or(default_instance))?;
+        if let Some(ref cs) = params.chip_select {
+            spi.set_pins(None, None, None, Some(&self.gpio_pin(cs.as_str())?))?;
+        }
+        if let Some(speed) = params.speed {
+            spi.set_max_speed(speed)?;
+        }
+        if let Some(voltage) = params.voltage {
+            spi.set_voltage(voltage)?;
+        }
+        if let Some(mode) = params.mode {
+            spi.set_transfer_mode(mode)?;
+        }
+        Ok(spi)
+    }
+
+    pub fn create_i2c(
+        &self,
+        params: &crate::io::i2c::I2cParams,
+        default_instance: &str,
+    ) -> Result<Rc<dyn Bus>> {
+        let i2c = self.i2c(params.bus.as_deref().unwrap_or(default_instance))?;
+        if let Some(speed) = params.speed {
+            i2c.set_max_speed(speed)?;
+        }
+        if let Some(addr) = params.addr {
+            i2c.set_default_address(addr)?;
+        }
+        Ok(i2c)
+    }
+
+    pub fn create_jtag<'t>(&'t self, params: &JtagParams) -> Result<Box<dyn JtagChain + 't>> {
+        let jtag = self.jtag(params)?;
+        Ok(jtag)
+    }
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
