+++ sw/host/opentitanlib/src/bootstrap/mod.rs
@@ -3,13 +3,12 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::rc::Rc;
-use std::time::Duration;
 
 use anyhow::Result;
-use clap::{Args, ValueEnum};
-use humantime::parse_duration;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
+
+pub use crate::io::{BootstrapOptions, BootstrapProtocol};
 
 use crate::app::{NoProgressBar, TransportWrapper, UartRx};
 use crate::impl_serializable_error;
@@ -33,22 +32,6 @@ pub enum BootstrapError {
 }
 impl_serializable_error!(BootstrapError);
 
-/// `BootstrapProtocol` describes the supported types of bootstrap.
-/// The `Primitive` SPI protocol is used by OpenTitan during development.
-/// The `Legacy` SPI protocol is used by previous generations of Google Titan-class chips.
-/// The `LegacyRescue` UART protocol is used by previous generations of Google Titan-class chips.
-/// The `Eeprom` SPI protocol is planned to be implemented for OpenTitan.
-/// The 'Emulator' value indicates that this tool has a direct way
-/// of communicating with the OpenTitan emulator, to replace the
-/// contents of the emulated flash storage.
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq, ValueEnum)]
-pub enum BootstrapProtocol {
-    Primitive,
-    Legacy,
-    LegacyRescue,
-    Eeprom,
-    Emulator,
-}
 
 // Implementations of bootstrap need to implement the `UpdateProtocol` trait.
 trait UpdateProtocol {
@@ -72,34 +55,7 @@ trait UpdateProtocol {
     ) -> Result<()>;
 }
 
-/// Options which control bootstrap behavior.
-/// The meaning of each of these values depends on the specific bootstrap protocol being used.
-#[derive(Clone, Debug, Args, Serialize, Deserialize)]
-pub struct BootstrapOptions {
-    #[command(flatten)]
-    pub uart_params: UartParams,
-    #[command(flatten)]
-    pub spi_params: SpiParams,
-    /// Bootstrap protocol to use.
-    #[arg(short, long, value_enum, ignore_case = true, default_value = "eeprom")]
-    pub protocol: BootstrapProtocol,
-    /// Whether to reset target and clear UART RX buffer after bootstrap. For Chip Whisperer board only.
-    #[arg(long)]
-    pub clear_uart: Option<bool>,
-    /// If set, keep the bootstrap strapping applied and do not perform the post-bootstrap reset
-    /// sequence.
-    #[arg(long)]
-    pub leave_in_bootstrap: bool,
-    /// If set, leave the reset signal asserted after completed bootstrapping.
-    #[arg(long)]
-    pub leave_in_reset: bool,
-    /// Duration of the inter-frame delay.
-    #[arg(long, value_parser = parse_duration)]
-    pub inter_frame_delay: Option<Duration>,
-    /// Duration of the flash-erase delay.
-    #[arg(long, value_parser = parse_duration)]
-    pub flash_erase_delay: Option<Duration>,
-}
+
 
 /// Bootstrap wraps and drives the various bootstrap protocols.
 pub struct Bootstrap<'a> {
