+++ sw/host/opentitanlib/src/io/mod.rs
@@ -2,14 +2,24 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod bootstrap_types;
+pub mod capabilities;
 pub mod console;
 pub mod eeprom;
+
 pub mod emu;
+pub mod errors;
 pub mod gpio;
+
+pub use bootstrap_types::{BootstrapOptions, BootstrapProtocol};
+pub use capabilities::{Capabilities, Capability, NeededCapabilities};
+pub use errors::{TransportError, TransportInterfaceType};
+pub use serializable_error::SerializableError;
 pub mod i2c;
 pub mod ioexpander;
 pub mod jtag;
 pub mod nonblocking_help;
+pub mod serializable_error;
 pub mod spi;
 pub mod uart;
 pub mod usb;
