+++ sw/host/opentitanlib/src/proxy/errors.rs
@@ -2,31 +2,7 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-use erased_serde::Serialize;
-
-/// `SerializableError` is a trait which represents an error type that can
-/// be sent over the wire by the proxy protocol.
-#[typetag::serde(tag = "error_type")]
-pub trait SerializableError: Serialize + std::error::Error + std::fmt::Debug + Send + Sync {
-    fn as_anyhow_error(self: Box<Self>) -> anyhow::Error;
-}
-
-/// `impl_serializable_error` needs to be invoked for every error type
-/// that can be sent over the wire.
-#[macro_export]
-macro_rules! impl_serializable_error {
-    ($t:ty) => {
-        const _: () = {
-            use $crate::proxy::errors::SerializableError;
-            #[typetag::serde]
-            impl SerializableError for $t {
-                fn as_anyhow_error(self: Box<$t>) -> anyhow::Error {
-                    self.into()
-                }
-            }
-        };
-    };
-}
+use crate::io::SerializableError;
 
 /// `SerializedError` is the wire form of errors that can be sent through the proxy.
 #[derive(serde::Serialize, serde::Deserialize, Debug)]
