+++ sw/host/opentitanlib/src/test_utils/init.rs
@@ -149,7 +149,7 @@ impl InitializeTest {
         }
 
         // Create the UART first to initialize the desired parameters.
-        let _uart = self.bootstrap.options.uart_params.create(&transport)?;
+        let _uart = transport.create_uart(&self.bootstrap.options.uart_params)?;
 
         // Load a bitstream.
         Self::print_result("load_bitstream", self.load_bitstream.init(&transport))?;
