+++ sw/host/opentitantool/src/command/console.rs
@@ -90,7 +90,7 @@ impl CommandDispatch for Console {
             };
             let mut stdout = std::io::stdout();
 
-            let uart = self.params.create(transport)?;
+            let uart = transport.create_uart(&self.params)?;
             if let Some(send) = self.send.as_ref() {
                 log::info!("Sending: {:?}", send);
                 uart.write(send.as_bytes())?;
